loader: turn script into a module, work in progress
