+++ warehouse-loader/warehouse/components/constants.py
@@ -1,0 +1,14 @@
+TRAINING_PREFIX = "training/"
+VALIDATION_PREFIX = "validation/"
+CONFIG_KEY = "config.json"
+
+TRAINING_PERCENTAGE = 0
+
+MODALITY = {
+    "DX": "xray",
+    "CR": "xray",
+    "MR": "mri",
+    "CT": "ct",
+}
+
+KB = 1024
